+++ src/blocking/request.rs
@@ -3,16 +3,12 @@ use std::fmt;
 use std::time::Duration;
 
 use http::{request::Parts, Request as HttpRequest, Version};
-use serde::Serialize;
-#[cfg(feature = "json")]
-use serde_json;
-use serde_urlencoded;
 
 use super::body::{self, Body};
 #[cfg(feature = "multipart")]
 use super::multipart;
 use super::Client;
-use crate::header::{HeaderMap, HeaderName, HeaderValue, CONTENT_TYPE};
+use crate::header::{HeaderMap, HeaderName, HeaderValue};
 use crate::{async_impl, Method, Url};
 
 /// A request which can be executed with `Client::execute()`.
@@ -391,10 +387,16 @@ impl RequestBuilder {
     /// as `.query(&[("key", "val")])`. It's also possible to serialize structs
     /// and maps into a key-value pair.
     ///
+    /// # Optional
+    ///
+    /// This requires the optional `query` feature enabled.
+    ///
     /// # Errors
     /// This method will fail if the object you provide cannot be serialized
     /// into a query string.
-    pub fn query<T: Serialize + ?Sized>(mut self, query: &T) -> RequestBuilder {
+    #[cfg(feature = "query")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "query")))]
+    pub fn query<T: serde::Serialize + ?Sized>(mut self, query: &T) -> RequestBuilder {
         let mut error = None;
         if let Ok(ref mut req) = self.request {
             let url = req.url_mut();
@@ -446,17 +448,23 @@ impl RequestBuilder {
     /// # }
     /// ```
     ///
+    /// # Optional
+    ///
+    /// This requires the optional `form` feature enabled.
+    ///
     /// # Errors
     ///
     /// This method fails if the passed value cannot be serialized into
     /// url encoded format
-    pub fn form<T: Serialize + ?Sized>(mut self, form: &T) -> RequestBuilder {
+    #[cfg(feature = "form")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "form")))]
+    pub fn form<T: serde::Serialize + ?Sized>(mut self, form: &T) -> RequestBuilder {
         let mut error = None;
         if let Ok(ref mut req) = self.request {
             match serde_urlencoded::to_string(form) {
                 Ok(body) => {
                     req.headers_mut()
-                        .entry(CONTENT_TYPE)
+                        .entry(crate::header::CONTENT_TYPE)
                         .or_insert(HeaderValue::from_static(
                             "application/x-www-form-urlencoded",
                         ));
@@ -504,14 +512,16 @@ impl RequestBuilder {
     /// fail, or if `T` contains a map with non-string keys.
     #[cfg(feature = "json")]
     #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    pub fn json<T: Serialize + ?Sized>(mut self, json: &T) -> RequestBuilder {
+    pub fn json<T: serde::Serialize + ?Sized>(mut self, json: &T) -> RequestBuilder {
         let mut error = None;
         if let Ok(ref mut req) = self.request {
             match serde_json::to_vec(json) {
                 Ok(body) => {
-                    if !req.headers().contains_key(CONTENT_TYPE) {
-                        req.headers_mut()
-                            .insert(CONTENT_TYPE, HeaderValue::from_static("application/json"));
+                    if !req.headers().contains_key(crate::header::CONTENT_TYPE) {
+                        req.headers_mut().insert(
+                            crate::header::CONTENT_TYPE,
+                            HeaderValue::from_static("application/json"),
+                        );
                     }
                     *req.body_mut() = Some(body.into());
                 }
@@ -547,7 +557,7 @@ impl RequestBuilder {
     #[cfg_attr(docsrs, doc(cfg(feature = "multipart")))]
     pub fn multipart(self, mut multipart: multipart::Form) -> RequestBuilder {
         let mut builder = self.header(
-            CONTENT_TYPE,
+            crate::header::CONTENT_TYPE,
             format!("multipart/form-data; boundary={}", multipart.boundary()).as_str(),
         );
         if let Ok(ref mut req) = builder.request {
@@ -683,13 +693,8 @@ fn fmt_request_fields<'a, 'b>(
 mod tests {
     use super::super::{body, Client};
     use super::{HttpRequest, Request, Version};
-    use crate::header::{HeaderMap, HeaderValue, ACCEPT, CONTENT_TYPE, HOST};
+    use crate::header::{HeaderMap, HeaderValue, ACCEPT, HOST};
     use crate::Method;
-    use serde::Serialize;
-    #[cfg(feature = "json")]
-    use serde_json;
-    use serde_urlencoded;
-    use std::collections::{BTreeMap, HashMap};
     use std::time::Duration;
 
     #[test]
@@ -825,6 +830,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "query")]
     fn add_query_append() {
         let client = Client::new();
         let some_url = "https://google.com/";
@@ -838,6 +844,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "query")]
     fn add_query_append_same() {
         let client = Client::new();
         let some_url = "https://google.com/";
@@ -850,8 +857,9 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "query")]
     fn add_query_struct() {
-        #[derive(Serialize)]
+        #[derive(serde::Serialize)]
         struct Params {
             foo: String,
             qux: i32,
@@ -873,8 +881,9 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "query")]
     fn add_query_map() {
-        let mut params = BTreeMap::new();
+        let mut params = std::collections::BTreeMap::new();
         params.insert("foo", "bar");
         params.insert("qux", "three");
 
@@ -889,19 +898,20 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "form")]
     fn add_form() {
         let client = Client::new();
         let some_url = "https://google.com/";
         let r = client.post(some_url);
 
-        let mut form_data = HashMap::new();
+        let mut form_data = std::collections::HashMap::new();
         form_data.insert("foo", "bar");
 
         let mut r = r.form(&form_data).build().unwrap();
 
         // Make sure the content type was set
         assert_eq!(
-            r.headers().get(CONTENT_TYPE).unwrap(),
+            r.headers().get(crate::header::CONTENT_TYPE).unwrap(),
             &"application/x-www-form-urlencoded"
         );
 
@@ -918,13 +928,16 @@ mod tests {
         let some_url = "https://google.com/";
         let r = client.post(some_url);
 
-        let mut json_data = HashMap::new();
+        let mut json_data = std::collections::HashMap::new();
         json_data.insert("foo", "bar");
 
         let mut r = r.json(&json_data).build().unwrap();
 
         // Make sure the content type was set
-        assert_eq!(r.headers().get(CONTENT_TYPE).unwrap(), &"application/json");
+        assert_eq!(
+            r.headers().get(crate::header::CONTENT_TYPE).unwrap(),
+            &"application/json"
+        );
 
         let buf = body::read_to_string(r.body_mut().take().unwrap()).unwrap();
 
@@ -983,6 +996,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "query")]
     fn normalize_empty_query() {
         let client = Client::new();
         let some_url = "https://google.com/";