+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.12.23
 
+- Add `query` and `form` features to make `serde` dependency optional.
 - Add `ClientBuilder::unix_socket(path)` option that will force all requests over that Unix Domain Socket.
 - Add `ClientBuilder::retry(policy)` and `reqwest::retry::Builder` to configure automatic retries.
 - Add `ClientBuilder::dns_resolver2()` with more ergonomic argument bounds, allowing more resolver implementations.